API/requirements.txt
mysql-connector-python==8.0.29
pyjwt==1.7.1